+++ src-dbg/flow_storm/debugger/events_processor.clj
@@ -108,7 +108,8 @@
     (ui-utils/run-later
       (flows-screen/update-exceptions-combo flow-id)
       ;; the first time we encounter an exception, navigate to that location
-      (when (= 1 (count (dbg-state/flow-exceptions flow-id)))
+      (when (and (:auto-jump-on-exception? (dbg-state/debugger-config))
+                 (= 1 (count (dbg-state/flow-exceptions flow-id))))
         (flows-screen/goto-location unwind-data)))))
 
 (defn data-window-push-val-data-event [{:keys [dw-id val-data root?]}]